pyproject.toml
[project]
name = "pryva"
version = "0.1.0"
description = "A secure command-line password manager"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "cryptography>=42.0.0",
    "click>=8.0.0",
    "argon2-cffi>=23.0.0",
]

[project.optional-dependencies]
clipboard = ["pyperclip>=1.8.0"]
test = ["pytest>=7.0.0"]

[project.scripts]
pryva = "pryva.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
